Chore/refactor react components (#55)
* Unused

* Moved componentes to their own files
