+++ web/src/HelmReleases.jsx
@@ -1,0 +1,37 @@
+import React, { useMemo, useState } from 'react';
+import { HelmRelease } from "./HelmRelease"
+import { filterResources } from './utils.js';
+
+export function HelmReleases(props) {
+  const { capacitorClient, helmReleases, targetReference, handleNavigationSelect } = props
+  const [filter, setFilter] = useState(false)
+  const sortedHelmReleases = useMemo(() => {
+    if (!helmReleases) {
+      return null;
+    }
+
+    return [...helmReleases].sort((a, b) => a.metadata.name.localeCompare(b.metadata.name));
+  }, [helmReleases]);
+
+  const filteredHelmReleases = filterResources(sortedHelmReleases, filter)
+
+  return (
+    <div className="space-y-4">
+      <button className={(filter ? "text-blue-50 bg-blue-600" : "bg-gray-50 text-gray-600") + " rounded-full px-3"}
+        onClick={() => setFilter(!filter)}
+      >
+        Filter errors
+      </button>
+      {
+        filteredHelmReleases?.map(helmRelease =>
+          <HelmRelease
+            key={"hr-" + helmRelease.metadata.namespace + helmRelease.metadata.name}
+            capacitorClient={capacitorClient}
+            item={helmRelease}
+            handleNavigationSelect={handleNavigationSelect}
+            targetReference={targetReference}
+          />
+        )}
+    </div>
+  )
+}
